+++ state-chain/pallets/cf-elections/src/electoral_systems.rs
@@ -6,7 +6,7 @@ pub mod liveness;
 pub mod mock;
 pub mod monotonic_change;
 pub mod monotonic_median;
-pub mod solana_swap_accounts_tracking;
+pub mod solana_vault_swap_accounts;
 pub mod unsafe_median;
 
 #[cfg(test)]
